// src/types/messages.ts
export interface GetStateMessage {
  command: 'getState';
}

export interface SetTrajectoryFrameMessage {
  command: 'setTrajectoryFrame';
  frameIndex: number;
}

export interface BeginDragMessage {
  command: 'beginDrag';
  atomId: string;
}

export interface EndDragMessage {
  command: 'endDrag';
}

export interface UndoMessage {
  command: 'undo';
}

export interface RedoMessage {
  command: 'redo';
}

export interface SelectAtomMessage {
  command: 'selectAtom';
  atomId: string;
  add?: boolean;
}

export interface SetSelectionMessage {
  command: 'setSelection';
  atomIds: string[];
}

export interface SelectBondMessage {
  command: 'selectBond';
  bondKey?: string;
  add?: boolean;
}

export interface SetBondSelectionMessage {
  command: 'setBondSelection';
  bondKeys: string[];
}

export interface ToggleUnitCellMessage {
  command: 'toggleUnitCell';
}

export interface SetUnitCellMessage {
  command: 'setUnitCell';
  params: {
    a: number;
    b: number;
    c: number;
    alpha: number;
    beta: number;
    gamma: number;
  };
  scaleAtoms?: boolean;
}

export interface ClearUnitCellMessage {
  command: 'clearUnitCell';
}

export interface CenterToUnitCellMessage {
  command: 'centerToUnitCell';
}

export interface SetSupercellMessage {
  command: 'setSupercell';
  supercell: [number, number, number];
}

export interface AddAtomMessage {
  command: 'addAtom';
  element: string;
  x: number;
  y: number;
  z: number;
}

export interface DeleteAtomMessage {
  command: 'deleteAtom';
  atomId: string;
}

export interface DeleteAtomsMessage {
  command: 'deleteAtoms';
  atomIds: string[];
}

export interface MoveAtomMessage {
  command: 'moveAtom';
  atomId: string;
  x: number;
  y: number;
  z: number;
  preview?: boolean;
}

export interface MoveGroupMessage {
  command: 'moveGroup';
  atomIds: string[];
  dx: number;
  dy: number;
  dz: number;
  preview?: boolean;
}

export interface SetAtomsPositionsMessage {
  command: 'setAtomsPositions';
  atomPositions: Array<{
    id: string;
    x: number;
    y: number;
    z: number;
  }>;
  preview?: boolean;
}

export interface EndDragEditMessage {
  command: 'endDrag';
}

export interface CopyAtomsMessage {
  command: 'copyAtoms';
  atomIds: string[];
  offset: {
    x: number;
    y: number;
    z: number;
  };
}

export interface ChangeAtomsMessage {
  command: 'changeAtoms';
  atomIds: string[];
  element: string;
}

export interface SetAtomColorMessage {
  command: 'setAtomColor';
  atomIds: string[];
  color: string;
}

export interface UpdateAtomMessage {
  command: 'updateAtom';
  atomId: string;
  element?: string;
  x?: number;
  y?: number;
  z?: number;
}

export interface SetBondLengthMessage {
  command: 'setBondLength';
  atomIds: string[];
  length: number;
}

export interface CreateBondMessage {
  command: 'createBond';
  atomIds: string[];
}

export interface DeleteBondMessage {
  command: 'deleteBond';
  bondKey?: string;
  atomIds?: string[];
  bondKeys?: string[];
}

export interface RecalculateBondsMessage {
  command: 'recalculateBonds';
}

export interface SaveStructureMessage {
  command: 'saveStructure';
}

export interface SaveStructureAsMessage {
  command: 'saveStructureAs';
}

export interface SaveRenderedImageMessage {
  command: 'saveRenderedImage';
  dataUrl: string;
  suggestedName?: string;
  width?: number;
  height?: number;
}

export interface OpenSourceMessage {
  command: 'openSource';
}

export interface ReloadStructureMessage {
  command: 'reloadStructure';
}

export interface GetDisplayConfigsMessage {
  command: 'getDisplayConfigs';
}

export interface LoadDisplayConfigMessage {
  command: 'loadDisplayConfig';
  configId: string;
}

export interface PromptSaveDisplayConfigMessage {
  command: 'promptSaveDisplayConfig';
  settings?: DisplaySettings;
}

export interface SaveDisplayConfigMessage {
  command: 'saveDisplayConfig';
  name: string;
  settings: DisplaySettings;
  description?: string;
  existingId?: string;
}

export interface GetCurrentDisplaySettingsMessage {
  command: 'getCurrentDisplaySettings';
}

export interface UpdateDisplaySettingsMessage {
  command: 'updateDisplaySettings';
  settings: DisplaySettings;
}

export interface ExportDisplayConfigsMessage {
  command: 'exportDisplayConfigs';
}

export interface ImportDisplayConfigsMessage {
  command: 'importDisplayConfigs';
}

export interface ConfirmDeleteDisplayConfigMessage {
  command: 'confirmDeleteDisplayConfig';
  configId: string;
}

export interface DeleteDisplayConfigMessage {
  command: 'deleteDisplayConfig';
  configId: string;
}

/**
 * All messages sent from webview to extension
 */
export type WebviewToExtensionMessage =
  | GetStateMessage
  | SetTrajectoryFrameMessage
  | BeginDragMessage
  | EndDragMessage
  | UndoMessage
  | RedoMessage
  | SelectAtomMessage
  | SetSelectionMessage
  | SelectBondMessage
  | SetBondSelectionMessage
  | ToggleUnitCellMessage
  | SetUnitCellMessage
  | ClearUnitCellMessage
  | CenterToUnitCellMessage
  | SetSupercellMessage
  | AddAtomMessage
  | DeleteAtomMessage
  | DeleteAtomsMessage
  | MoveAtomMessage
  | MoveGroupMessage
  | SetAtomsPositionsMessage
  | EndDragEditMessage
  | CopyAtomsMessage
  | ChangeAtomsMessage
  | SetAtomColorMessage
  | UpdateAtomMessage
  | SetBondLengthMessage
  | CreateBondMessage
  | DeleteBondMessage
  | RecalculateBondsMessage
  | SaveStructureMessage
  | SaveStructureAsMessage
  | SaveRenderedImageMessage
  | OpenSourceMessage
  | ReloadStructureMessage
  | GetDisplayConfigsMessage
  | LoadDisplayConfigMessage
  | PromptSaveDisplayConfigMessage
  | SaveDisplayConfigMessage
  | GetCurrentDisplaySettingsMessage
  | UpdateDisplaySettingsMessage
  | ExportDisplayConfigsMessage
  | ImportDisplayConfigsMessage
  | ConfirmDeleteDisplayConfigMessage
  | DeleteDisplayConfigMessage;